:sparkles: add type annotation
