+++ deepchem/metrics/tests/test_genomics.py
@@ -2,17 +2,16 @@
 Test that genomic metrics work.
 """
 import unittest
-import os
 
 import numpy as np
 import deepchem as dc
 import tensorflow as tf
 
-LETTERS = "ACGT"
-
 from deepchem.metrics.genomic_metrics import get_motif_scores
 from deepchem.metrics.genomic_metrics import get_pssm_scores
 from deepchem.metrics.genomic_metrics import in_silico_mutagenesis
+
+LETTERS = "ACGT"
 
 
 class TestGenomicMetrics(unittest.TestCase):
@@ -35,7 +34,6 @@ class TestGenomicMetrics(unittest.TestCase):
 
   def test_get_pssm_scores(self):
     """Test get_pssm_scores returns correct shape."""
-    motif_name = "TAL1_known4"
     sequences = np.array(["ACGTA", "GATAG", "CGCGC"])
     sequences = dc.utils.genomics_utils.seq_one_hot_encode(
         sequences, letters=LETTERS)
@@ -66,9 +64,6 @@ class TestGenomicMetrics(unittest.TestCase):
     labels = np.reshape(labels, (3, 1))
     self.assertEqual(sequences.shape, (3, 4, 5, 1))
 
-    #X = np.random.rand(10, 1, 4, 50)
-    #y = np.random.randint(0, 2, size=(10, 1))
-    #dataset = dc.data.NumpyDataset(X, y)
     dataset = dc.data.NumpyDataset(sequences, labels)
     model = self.create_model_for_mutagenesis()
     model.fit(dataset, nb_epoch=1)
@@ -87,9 +82,6 @@ class TestGenomicMetrics(unittest.TestCase):
     labels = np.reshape(labels, (3, 1))
     self.assertEqual(sequences.shape, (3, 4, 5, 1))
 
-    #X = np.random.rand(10, 1, 4, 50)
-    #y = np.random.randint(0, 2, size=(10, 1))
-    #dataset = dc.data.NumpyDataset(X, y)
     dataset = dc.data.NumpyDataset(sequences, labels)
     model = self.create_model_for_mutagenesis()
     model.fit(dataset, nb_epoch=1)
